0.1.24 - compliance standard result uint8 -> uint256

+++ contracts/library/compliance/ComplianceStandard.sol
@@ -7,39 +7,39 @@ import "../provider/Provider.sol";
  * @dev An on-chain Compliance Provider.
  */
 contract ComplianceStandard is Provider {
-  /**
-   * @dev Checks to see if an action is compliant.
-   *
-   * @param instrumentAddr The address of the instrument contract.
-   * @param instrumentIdOrAmt The instrument id (NFT) or amount (ERC20).
-   * @param from The from address of the token transfer.
-   * @param to The to address of the token transfer.
-   * @param data Any additional data related to the action.
-   *
-   * @return result and next serviceId.
-   */
-  function performCheck(
-    address instrumentAddr,
-    uint256 instrumentIdOrAmt,
-    address from,
-    address to,
-    bytes32 data
-  ) view external returns (uint8, uint256);
+    /**
+    * @dev Checks to see if an action is compliant.
+    *
+    * @param instrumentAddr The address of the instrument contract.
+    * @param instrumentIdOrAmt The instrument id (NFT) or amount (ERC20).
+    * @param from The from address of the token transfer.
+    * @param to The to address of the token transfer.
+    * @param data Any additional data related to the action.
+    *
+    * @return result and next serviceId.
+    */
+    function performCheck(
+        address instrumentAddr,
+        uint256 instrumentIdOrAmt,
+        address from,
+        address to,
+        bytes32 data
+    ) view external returns (uint256, uint256);
 
-  /**
-   * @dev Called when a hard check is performed.
-   *
-   * @param instrumentAddr The address of the instrument contract.
-   * @param instrumentIdOrAmt The instrument id (NFT) or amount (ERC20).
-   * @param from The from address of the token transfer.
-   * @param to The to address of the token transfer.
-   * @param data Any additional data related to the action.
-   */
-  function performHardCheck(
-    address instrumentAddr,
-    uint256 instrumentIdOrAmt,
-    address from,
-    address to,
-    bytes32 data
-  ) external;
+    /**
+    * @dev Called when a hard check is performed.
+    *
+    * @param instrumentAddr The address of the instrument contract.
+    * @param instrumentIdOrAmt The instrument id (NFT) or amount (ERC20).
+    * @param from The from address of the token transfer.
+    * @param to The to address of the token transfer.
+    * @param data Any additional data related to the action.
+    */
+    function performHardCheck(
+        address instrumentAddr,
+        uint256 instrumentIdOrAmt,
+        address from,
+        address to,
+        bytes32 data
+    ) external;
 }
