+++ contracts/library/instruments/AbacusERC20Token.sol
@@ -19,12 +19,12 @@ contract AbacusERC20Token is StandardToken {
         complianceProviderId = _complianceProviderId;
     }
 
-    function canTransfer(address to, uint256 value) public view returns (uint8, uint256) {
+    function canTransfer(address to, uint256 value) public view returns (uint256, uint256) {
         return complianceCoordinator.check(complianceProviderId, this, value, msg.sender, to, 0);
     }
 
     function transfer(address to, uint256 value) public returns (bool) {
-        uint8 checkResult;
+        uint256 checkResult;
         (checkResult,) = complianceCoordinator.hardCheck(complianceProviderId, this, value, msg.sender, to, 0);
         if (checkResult != 0) {
             return false;
@@ -32,12 +32,12 @@ contract AbacusERC20Token is StandardToken {
         return super.transfer(to, value);
     }
 
-    function canTransferFrom(address from, address to, uint256 value) public view returns (uint8, uint256) {
+    function canTransferFrom(address from, address to, uint256 value) public view returns (uint256, uint256) {
         return complianceCoordinator.check(complianceProviderId, this, value, from, to, 0);
     }
 
     function transferFrom(address from, address to, uint256 value) public returns (bool) {
-        uint8 checkResult;
+        uint256 checkResult;
         (checkResult,) = complianceCoordinator.hardCheck(complianceProviderId, this, value, from, to, 0);
         if (checkResult != 0) {
             return false;
